requirements.txt
asgiref==3.4.1
bidict==0.21.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==1.16.0
eventlet==0.30.2
Flask==2.0.2
Flask-PyMongo==2.3.0
Flask-SocketIO==5.1.1
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonfield==3.1.0
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysqlclient==2.1.0
oauthlib==3.1.1
pdf2image==1.16.0
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
pymongo==3.12.1
python-dotenv==0.19.2
python-engineio==4.3.0
python-socketio==5.5.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
simple-websocket==0.4.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0